@@ -1,92 +1,41 @@
 package com.devrezaur.main.model;
 
-import java.util.List;
-import javax.persistence.CascadeType;
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.FetchType;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.JoinColumn;
-import javax.persistence.JoinTable;
-import javax.persistence.ManyToMany;
-import javax.persistence.Table;
-import org.springframework.stereotype.Component;
+import jakarta.persistence.*;
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
 
+import java.util.Set;
+
+@NoArgsConstructor
+@AllArgsConstructor
+@Data
 @Entity
-@Table(name = "users")
-@Component
+@Table(name = "user_table")
 public class User {
 
-	@Id
-	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	@Column(name = "user_id")
-	private long id;
+    @Id
+    @GeneratedValue
+    @Column(name = "user_id")
+    private long id;
 
-	@Column(name = "fullname")
-	private String fullname;
+    @Column(name = "first_name")
+    private String firstName;
 
-	@Column(name = "username", unique = true)
-	private String username;
+    @Column(name = "last_name")
+    private String lastName;
 
-	@Column(name = "password")
-	private String password;
+    @Column(name = "username", unique = true)
+    private String username;
 
-	@ManyToMany(cascade = CascadeType.MERGE, fetch = FetchType.EAGER)
-	@JoinTable(name = "user_role", joinColumns = @JoinColumn(name = "user_id"), inverseJoinColumns = @JoinColumn(name = "role_id"))
-	private List<Role> roles;
+    @Column(name = "password")
+    private String password;
 
-	public User() {
-
-	}
-
-	public User(long id, String fullname, String username, String password, List<Role> roles) {
-		this.id = id;
-		this.fullname = fullname;
-		this.username = username;
-		this.password = password;
-		this.roles = roles;
-	}
-
-	public long getId() {
-		return id;
-	}
-
-	public void setId(long id) {
-		this.id = id;
-	}
-
-	public String getFullname() {
-		return fullname;
-	}
-
-	public void setFullname(String fullname) {
-		this.fullname = fullname;
-	}
-
-	public String getUsername() {
-		return username;
-	}
-
-	public void setUsername(String username) {
-		this.username = username;
-	}
-
-	public String getPassword() {
-		return password;
-	}
-
-	public void setPassword(String password) {
-		this.password = password;
-	}
-
-	public List<Role> getRoles() {
-		return roles;
-	}
-
-	public void setRoles(List<Role> roles) {
-		this.roles = roles;
-	}
-
+    @ManyToMany(cascade = CascadeType.MERGE)
+    @JoinTable(
+            name = "user_authority",
+            joinColumns = @JoinColumn(name = "user_id"),
+            inverseJoinColumns = @JoinColumn(name = "authority_id")
+    )
+    private Set<Authority> authorities;
 }
