@@ -1,42 +1,44 @@
 package com.devrezaur.main.service;
 
-import java.util.Arrays;
-import org.springframework.beans.factory.annotation.Autowired;
+import com.devrezaur.main.model.Authority;
+import com.devrezaur.main.model.User;
+import com.devrezaur.main.repository.AuthorityRepository;
+import com.devrezaur.main.repository.UserRepository;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
-import com.devrezaur.main.model.Role;
-import com.devrezaur.main.model.User;
-import com.devrezaur.main.repository.RoleRepository;
-import com.devrezaur.main.repository.UserRepository;
+
+import java.util.Set;
 
 @Service
 public class UserService {
-	
-	@Autowired
-	private UserRepository userRepository;
-	@Autowired
-	private RoleRepository roleRepository;
-	@Autowired
-	private PasswordEncoder passwordEncoder;
 
-	public User findUserByUsername(String username) {
-		return userRepository.findByUsernameIgnoreCase(username);
-	}
+    private final UserRepository userRepository;
+    private final AuthorityRepository authorityRepository;
+    private final PasswordEncoder passwordEncoder;
 
-	public User saveUser(User user) {
-		user.setPassword(passwordEncoder.encode(user.getPassword()));
-		Role userRole = roleRepository.findByRole("ROLE_USER");
-		user.setRoles(Arrays.asList(userRole));
-		
-        return user = userRepository.save(user);
-	}
-	
-	public User saveAdmin(User user) {
-		user.setPassword(passwordEncoder.encode(user.getPassword()));
-		Role adminRole = roleRepository.findByRole("ROLE_ADMIN");
-		user.setRoles(Arrays.asList(adminRole));
-		
-        return user = userRepository.save(user);
-	}
+    public UserService(UserRepository userRepository, AuthorityRepository authorityRepository,
+                       PasswordEncoder passwordEncoder) {
+        this.userRepository = userRepository;
+        this.authorityRepository = authorityRepository;
+        this.passwordEncoder = passwordEncoder;
+    }
+
+    public User findUserByUsername(String username) {
+        return userRepository.findByUsernameIgnoreCase(username);
+    }
+
+    public void saveUser(User user) {
+        user.setPassword(passwordEncoder.encode(user.getPassword()));
+        Authority authority = authorityRepository.findByAuthority("ROLE_USER");
+        user.setAuthorities(Set.of(authority));
+        userRepository.save(user);
+    }
+
+    public void saveAdmin(User user) {
+        user.setPassword(passwordEncoder.encode(user.getPassword()));
+        Authority authority = authorityRepository.findByAuthority("ROLE_ADMIN");
+        user.setAuthorities(Set.of(authority));
+        userRepository.save(user);
+    }
 
 }
